pyproject.toml
[project]
name = "holes-in-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "streamlit>=1.51.0",
]
